docs, flag, image, pubkey: Allow downloading insecure GPG keys
When fetching images via a discovery mechanism, it is also possible to
fetch the keys via the discovery and then trust them to verify the
image signature. For the key to be fetched, HTTPS connection is
required and the server's certificate must be successfully
verified. Trusting the key puts it into a persistent keystore. This
makes it hard to test it with a local server, which usually doesn't
have a valid certificate.

The "pubkey" option of the --insecure-options flag makes fetching keys
from insecure places possible, that is - from HTTP connections or from
servers with an invalid certificate.