store: force removal of the tree store even if it's in a bad state
If the "image" file is not present in the tree store directory, that
tree store can't be removed. Don't error out if that happens, just print
a warning so it is removed.

This means that the image size will be overestimated because we can't
know which image the tree store refers to, but at least we can remove
the tree store.