doc: dependencies: linux 3.18
Linux 3.18 introduced the fix for unlinking files and directories that
are mount points in another mount namespace. Before Linux 3.18,
unlinking could return EBUSY:
https://github.com/torvalds/linux/commit/8ed936b

With the introduction of rkt-fly, this becomes a visible issue with rkt
gc, see discussion on:
https://github.com/coreos/rkt/issues/1922#issuecomment-196914006

This patch bumps the run-time dependency to Linux 3.18. This also
documents the issues in README.md.