stage1/gc: remove journal link
In the nspawn flavors, we use the "--try-guest" option. This means a
symlink is created from `/var/log/journal/${machine-id}` to pod's
`/var/log/journal/{machine-id}` when the pod is started.

If that symlink exists when doing GC, remove it.

If systemd on the host is configured to use `/run/log/journal/`, we
don't care about that because it will be deleted on a reboot.