stage1: implement docker volume semantics
Docker volumes are initialized with the files in the image if they
exist, unless a host directory is mounted there (see
https://docs.docker.com/engine/userguide/containers/dockervolumes/#data-volumes).

However, when we convert docker images with docker2aci, we convert
docker volumes to mount points, and if the user doesn't specify a host
path for mounting, we mount an empty volume, effectively masking the
files in the image.

To fix this, we create a special case for the case when application uses
an image converted from docker and we mount an implicit empty volume.
In that case, we copy the files from the image to the volume, so docker
images work correctly.

Note that we won't do this if the user explicitly specifies an empty
volume or if they mount a host directory.