build: Make sure we really got all the files from the squashfs file
Since unsquashfs does not have a flag for ensuring that all the files
from manifest were extracted, we have to do it on our own. We do this
by making sure that the complete manifest is a subset of a list of the
files in the squashfs file (acquired by the "unsquashfs -ls <file>"
command). This requires doing some set operations, so we got a
dependency on comm from coreutils.