rkt: add command to export a pod to an aci
This adds a new `export` command to rkt which generates an aci from a
pod; saving any changes made to the pod.

Usage is `rkt export POD_UUID my.aci`

This is currently restricted to only working with exited pods with a
single app and that have been run with the --no-overlay option.

Fixes #1197