functional tests: capture SIGTERM in inspect
In TestService, we start a systemd service running a rkt image that
prints a message every second. Then we stop the unit and check it's not
in the service list.

This was working before because we weren't propagating the exit status
of apps in the pod to the outside.

Now that we do it, stopping the unit file means that SIGTERM will be
sent to nspawn, which will trigger a shutdown of systemd inside the pod,
which means SIGTERM will be sent to all the apps and that makes the
inspect binary exit with a non-zero status. Transient units that exit
with a failed status remain on the services list, so our test fails.

Capture SIGTERM in inspect and return 0 in that case, so the test
passes.