manpages: generate manpages with cobra
To update the manpages, use:
    $ make manpages

It will generate the manpages in dist/manpages/.

Also the bash completion file is no longer built manually, but you can
use:
    $ make bash-completion

It will generate dist/bash_completion/rkt.bash.

The release scripts are updated to make sure that each release
contains an up-to-date version of the manpages and the bash completion
file.