stage1/prepare-app: avoid recursive bind-mounts on /sys
Starting a pod with 100 apps is really slow. This patch fixes the
slowness.

prepare-app recursively bind-mounts /sys from stage1 to the stage2
(apps' rootfs). "Recursive" means it includes all the cgroup mounts
because they are in /sys/fs/cgroup. Moreover, rkt bind mounts some
cgroup knob files in the cgroup filesystem for enabling the memory and
cpu isolator.

The number of cgroup bind mounts in stage1 is linear with the number of
apps: O(n)

The number of cgroup bind mounts in stage2 is quadratic with the number
of apps: O(n^2)

With one app, I have 17 bind mounts related to cgroups. With 100 apps,
17 * 100 * 100 = 170.000 bind mounts.

For each change in the mount table, systemd is notified via inotify on
/proc/self/mountinfo and it checks the configuration of that mount in
/etc/systemd/system, /run/systemd/system, /usr/local/lib/systemd/system
and /usr/lib64/systemd/system.

systemd does about 30 syscalls per new mount notified via
/proc/self/mountinfo. That was 5.100.000 syscalls for mounting
cgroups in a 100-app pod.

This patch changes the logic to bind mount /sys:
- in the usual non-unified cgroup hierarchy: bind mount /sys
  non-recursively and instead bind-mount each cgroup hierarchy
  individually, non-recursively.
- in the unified cgroup hierarchy: keep the recursive bind mount because
  we will not have per-app bind mounts on knob files in this case.