rkt/image: check that discovery labels match manifest labels
The spec in the discovery part says: "Implementations MUST ensure
that the initial name and labels used for discovery matches the name and
labels in the Image Manifest."

Actually rkt just checks that the names matches but ignore the labels.

This patch also check the labels. Like the other validation parts the
check is executed only if the user doesn't pass `--insecure-options=image`

A functional test is added.

With this patch users calling `rkt fetch example.com/aci:latest` but
receiving an aci with a different version (like v2.0) will receive a
validation error.