rkt: calculate real dataDir path
If dataDir or some of its parents are symlinks all seems to work correctly
except the gc.

That's because stage0.MountGC determines which mountpoints should be
unmounted filtering the mountpoints returned from /proc/self/mountinfo
with the dataDir value. These mountpoints are the real path so, if the
dataDir or its parents are symlinks, no mountpoint will match.

This patch evaluates the dataDir real path.