gc: Add flag 'mark-only' to mark garbage pods without deleting them.
If this flag is set to true, then we move the exited/aborted pods
to exited-garbage/garbage directory, but do not deleting them.

For prepared pods, if the flag is true, we don't move them to garbage
even the pods are expired.

A third party application can use 'rkt gc --mark-only=true' to
marks the time when the pods is not running.