stage1/init: get annotations from the ImageManifest
To detect if we're running an image that was converted from docker, we
were getting annotations from the RuntimeApp. This doesn't work for the
case where the user runs rkt directly with a pod manifest, since they
can omit the annotations there.

Instead, use the ImageManifest, which is the right place to look for
annotations.