stage1/init: create empty volume directory always
If there was an implicit empty volume and the image was converted from
docker, we copy the files from the image to the volume.

However, if the image defines the mountpoint but the directory is not
actually present in the image, CopyTree will do nothing and the volume
directory won't be created, making the following Chown call fail.

We now create the directory always. If CopyTree already created it, it
doesn't hurt.