net/portfwd: choose network according to ipmasq
When a network has ipmasq enabled, CNI generates masquerading rules for
the networks subnet. These rules are appended, so the first network is
the most upper one. Hence, rkt needs to use the address of first
masqueraded network as packets will always hit this subnet's CNI rule
first.

If no network has ipmasq enabled, we default to the previous behavior of
using the last loaded network, which is either the default or
default-restricted network.