stage1: always write /etc/machine-id
Prepare rkt for systemd-v230 in stage1.

rkt was previously only writing /etc/machine-id in the stage1 rootfs if
systemd was running on the host. However, new versions of systemd always
require /etc/machine-id, see:
https://github.com/systemd/systemd/commit/e01ff70a77e781734e1e73a2238af2e9bf7967a8
This patch will be in systemd-v230.

This is currently not a problem because stage1 does not use systemd-v230
(and systemd-v230 is not released yet), but we want rkt to be ready.

Fixes https://github.com/coreos/rkt/issues/2416