rkt/image: strip "Authorization" on redirects to a different host
We shouldn't pass the "Authorization" header if the redirect goes to a
different host, it can leak sensitive information to unexpected third
parties.

This fixes an issue where rkt gets redirected to Amazon S3 when
downloading an ACI from quay and S3 returns a 400 (Bad Request) with
this error message:

"Only one auth mechanism allowed; only the X-Amz-Algorithm query
parameter, Signature query string parameter or the Authorization header
should be specified"

The problem happens because we're sending "Authorization" and the URL on
the redirect already includes a "Signature" query string parameter.