stage1: use pure systemd to start apps in a pod
Up to this point, we were using appexec as a wrapper to the actual
binary to set up the environment for the apps in a pod. This commit
replaces appexec's functionality with pure systemd options:

appexec        | systemd option
------------   | -------------
chroot         | RootDirectory=
chdir          | WorkingDirectory=
setenv         | EnvironmentFile=
setresuid      | User=
setresgid      | Group=
setgroups      | SupplementaryGroups=
execvp         | ExecStart=
$PATH lookup   | implemented when writing the unit file
ELF diagnostic | not yet implemented

We're still using appexec (it'll be renamed in a follow-up commit) to
implement the `rkt enter` functionality.

Implementing this in systemd allows us to restrict the capabilities
granted to apps, including `CAP_SYS_ADMIN`.