stage1/prepare-app: check for user namespace env
Currently prepare-app always bind mounts /sys non-recursively if the
non-unified cgroups hierarchy is present to prevent O(n2) behavior
caused by recursive mounting.

This optimization fails if a container is started in a user namespace
environment where the kernel prevents from mounting /sys
non-recursively.

This checks explicitely if we are in a user namespaced environment and
falls back to a recursive bind mount in this case.

Fixes #2490