functional tests: don't use gnupg for signing images
Some distros only provide gnupg2 (while others provides both version)
and the `gpg` command is a symlink to `gpg2`.

With gnupg >= 2.1 signing test images fails due to changes in the
secring handling (see
https://www.gnupg.org/faq/whats-new-in-2.1.html#nosecring)

To make tests work with gnupg >= 2.1 the `runSignImage` function should be
changed to do these steps:

* Create a temporary gnupg home (with secure permissions like 700).
* Copy the secring.gpg file to that gnupg home.
* Call gpg with this command line: `gpg --homedir %s --default-key %s --output %s --detach-sig %s`

If the gpg command is provided by gnupg2 it'll convert the secring to
the format required by the gpg-agent in the `private-keys-v1.d`
directory.

Additionally gpg2 will spawn a gpg-agent to handle the private
keys and there'll be an additional problem when the path to the
gpg-agent socket files is too long (see
http://comments.gmane.org/gmane.comp.encryption.gpg.devel/21073), so
it'll require a gnupg home inside a short path.

To avoid all of this, since the functional tests only require gpg for
signing tests images this patch uses the `golang.org/x/crypto/openpgp`
package.

Fixes #2424.