rkt/image: render images only if we're root
To render images, we need access to the tree store directories and
capabilities to do a chroot.

We could allow the rkt group to access the tree store directories and
not do a chroot when rendering but since this rendering-on-fetch is just
an optimization, it's simpler to just skip the rendering if we fetch as
non-root.