lkvm: Bump the version of upstream repository
glibc >= 2.23 deprecated the readdir_r usage in favor
of readdir. Build of lkvm is treating warnings as errors,
so it's impossible to build rkt with kvm flawor with the
newest glibc.

That was fixed in the upstream commit[1].

Also, the another commit[2] fixed an issue with dead lock
on reboot, so dead_lock_fix patch isn't needed anymore.

[1] https://kernel.googlesource.com/pub/scm/linux/kernel/git/will/kvmtool/+/d62653e177597251c24494a6dda60acd6d846671
[2] https://kernel.googlesource.com/pub/scm/linux/kernel/git/will/kvmtool/+/e8cb90fb9697da24cdf553a1b8cf120f94c20832

Fixes #2594

Signed-off-by: Michal Rostecki <michal.rostecki@gmail.com>