stage1/init: work around cgroup/SCM_CREDENTIALS race
There's a race[1][1] on systemd that causes the systemd unit name not
getting written to the journal for short-lived non-root services.

To provide a way to identify an app within a pod in the logs, we set
`SyslogIdentifierp` to the app name as a workaround.

This causes processes forked by the app and pre/post start hooks to be
identified as the app, but it was already happening before we removed
appexec[2][2] with the binary exec'd by the app instead of the app name.

In theory we could remove that and we'll get the binary name of each
process executed in the pod as the syslog identifier, but then this
workaround would not *work*.

Setting the app name as the identifier is already an improvement over
the previous situation.

[1]: https://github.com/systemd/systemd/issues/2913
[2]: https://github.com/coreos/rkt/pull/2493