rkt: fix panic on rm a non-existing pod with uuid-file
We were trusting that the value in the uuid file referred to an existing
pod, which caused a panic if that wasn't the case.

In this commit we resolve the UUID to check if the pod actually exists.

Also, make rm's behavior consistent between caling it with a pod UUID
and with `--uuid-file`:

* Print a similar error message
* Set the exit status to 1 if we fail to remove any pod