build: Find also blackbox tests for unit-check
We were only checking the packages that had non-empty TestGoFiles
field returned by "go list". But the TestGoFiles contains only files
that have the same package as the non-test files in the same
directory. The test files with the different package name are in the
XTestGoFiles field, which we did not handle before.

This makes go test to run the test for pkg/selinux and pkg/user
packages too.