pkg/selinux: fix TestSELinux
TestSELinux was not setting the MCS dir which caused the paths of every
MCS to be in the root of the file system (the path is calculated as
fmt.Sprintf("%s/%s", mcsdir, mcs) and mcsdir was the empty string by
default). This was making the test to fail when it runs as non-root.

The test failure was non-obvious because of the way the algorithm to get
a unique MCS is implemented. We first try to create the MCS file with a
random category pair and, if it doesn't work, we continue with a
different random category pair. Since we can't write to the MCS
directory, we'll continue doing this until the test times out and we
don't print any errors when we can't create the MCS file.

Also, clean up test code a bit.