rkt.mk: Compile generators for *build* platform
manpages_gen and bash_completion_gen are compiled and run at build time,
so need to be built for the build platform, not the target platform.

This change overrides GOARCH/CC appropriately for these two targets.
Note building these tools requires that the systemd and acl C headers
are also available for the build architecture.