functional tests: skip test if a remote fetching fails
It was provoking so many false positives.

This means we won't always be doing a full TestFetch but in some
distributions the test will run until the end so we can verify we didn't
break anything.

Later we can consider retrying if a remote fetch fails. Or even better,
always fetch from the local test ACI server and implement/run our own
Docker server and fetch from that. Or wait for OCI support and implement
a local OCI server :).