functional tests: start api service with "nobody" UID
We now check if we're registered with machined over dbus before
returning from the ListPods API call. It turns out dbus authentication
doesn't succeed unless our UID is mapped to a user name, and UID 65534
might not be mapped to any user name.

Instead, look for the "nobody" user in `/etc/passwd` and use that. If
it's not found, fall back to running the API service as root in the
functional tests.