stage1_fly: write pid instead of ppid
Currently the stage1 fly flavor writes a ppid file. This implies it will
have a child pid eventually. For the fly flavor this assumption does not
hold. fly may or may not spawn children hence a ppid doesn't make sense.

This fixes it and writes a pid file instead.

Fixes partially #2792