functional tests: sleep before killing rkt on TestResumedFetch
Our test server sends half of the file and then tells TestResumedFetch
via a channel that it should kill rkt and start another rkt process
to check if it resumes the download.

However, it may happen that rkt doesn't have time to write everything to
disk before it gets killed, in some cases it doesn't even have time to
write anything and the test fails.

To fix it, we sleep a bit before telling TestResumedFetch that it should
kill rkt.