rkt/image: when using heuristics to determine app type, stat first
At some point rkt will fall back to using heuristics to determine if the
provided image name is referring to an ACI on disk or if it's an image
name to perform discovery on / look in the store for. This commit
modifies the function that determines this to begin by stat'ing the
provided image name. If a file to that exact path exists on disk, that's
probably what the user wished to run.