common: remove unused GetImageIDs()
This patch removes apps.GetImageIds() method as it has no callers anymore.
This helper was added to pass list of image ids to the state0, but now we
pass list of apps in a pod instead.

Last user of the GetImageIDs() was removed in the 227fe3c commit
( Move pod registration from stage1 to stage0. by @eyakubovich ) and it is
not used anywhere anymore.

So we can remove it.