stage1: Implemented notifications from the app in the container to the host
After this patch, systemd on the host marks the container as "active"
when the init systemd in the container sends the ready message
instead of doing it after the container is created.
After this patch, rkt adds the switch `--notify-ready=yes` when it runs the pod
with systemd-nspawn.

Fixes: https://github.com/coreos/rkt/issues/1464

Warning:
If the application launched by rkt does not notify (with sd_notify())
that it is ready, then the application may be killed after a timeout.