pkg/tar: remove errwrap
Since this package can be used outside of rkt, there is no need to
obscure the return values by using errwrap. All of the internal
consumers of these functions already wrap the output anyway, so this is
actually redundant. For example, extractTarCommand() may return:

  error extracting tarball
    error extracting tarball
      no such file or directory

This change removes the outermost wrap:

  error extracting tarball
    no such file or directory