rkt/run: handle malformed environment files
This patch returns an error if the environment file contains a malformed
variable. For example, if a variable is unnamed (e.g. `=value`) or without
the equals sign (e.g. `VAR_1`). Empty variables may be set by `VAR_1=`.
Users may prefix with `#` or `;` those varibles to ignore. This patch
skips empty lines in the environment file.