store: decouple image store and tree store implementations.
The current image store and tree store are logically decoupled but
historically the implementation was done under an unique store object.

This patch splits the store in its two logical parts: an image store
(currently only appc ACIs) and a tree store.

This makes the code cleaner and will ease implementations of different kind
of image types for other specs (like OCI).

Future patches will cover the various TODOs:

* Migration to a less unfortunate path naming ("cas") and separated
paths for the store and the tree store. This needs to be seamless and
is quite complex (have to handle pods with an overlayfs mounted etc.)
Make the tree store accept an interface to an image store (this needs
* The removal of the current logic of saving the aci tree store sizes
inside the aci store totally decoupling them and make `rkt image list`
calculate the total size)