namefetcher: Honor http Cache-Control max-age
The fetcher refactor in #1763 removed the http caching checks but left
them only in the http fetcher.

This patch adds the same logic also to the namefetcher and extends the
functional test to also check fetching by appc name.

In the meantime (it won't affect anything since there's no way to
currently ask for it) change the meaning of Fetcher.NoCache to not disable both store and
transport cache but only disable transport cache. This is a more logical
one and decoupled from the store as it should be.