functional tests: skip TestSocketProxyd
This test is currently racy. The socket-activated proxyd service depends
on a rkt service being ready. However, the application that runs inside
rkt doesn't have a way to signal when it's ready because of
https://github.com/coreos/rkt/issues/1464. Instead, systemd assumes it's
ready when the container starts, which is wrong, because the application
inside might not be listening on the port yet. This makes the test fail.

It fails a lot particularly in Fedora Rawhide with the flavor host:

https://jenkins-rkt-public.prod.coreos.systems/job/rkt-master-periodic/os_type=fedora-rawhide,stage1_flavor=host/

We can re-enable it when we fix https://github.com/coreos/rkt/issues/1464.