build: Make generating clean files simpler to do
This adds two functions. One takes a filelist and a directory map and
generates the clean files based on them. The other one takes the
directory, dependencies and the directory map, generates the filelist
based on a directory after the passed dependencies were met and then
generates the clean file based on the generated filelist and the
directory map.

The former is useful for just removing some directory with contents
generated during build. The latter is useful when we want to generate
the filelist ourselves, because the filelist will be used also for
something else (like defining dependencies).