stage1: shutdown.service: don't use /dev/console
The default stdout for shutdown.service is /dev/console (the tty created
by nspawn). But that tty might be destroyed if rkt is executed via ssh
and the user terminates the ssh session. We still want shutdown.service
to succeed in that case, so don't use /dev/console.

Fixes https://github.com/coreos/rkt/issues/2232