rkt: Implement distribution points.
This patch implements the distribution point concept with the three initial
kind of distributions already used by rkt (appc, aci-archive, docker).

It just makes the smallest required changes to the fetchers logic to make them
work and pass the functional tests.

This is the base for future fetchers (a future patch will refactor the
fetcher logic to better fit the distribution concept) and store refactors.