cri: correct 'created' timestamp
Prior to this change, the mtime of the 'pod' file is used for the created
timestamp.

This is inaccurate in the case of CRI where the pod manifest is changed
with each app added.

To resolve that problem, a new file is introduced which is not touched
by the app add codepath.