stage1/kvm: only configure build qemu when we really need to
 QEMU was constantly re-configured and built for every make. This
is not necessary. The root cause was the CONF stamp stage, which
was the target for the git clone and then configure steps, but had
no rule to actually generate the stamp file, so made the configure
stage think we had always just cloned (in effect).
 Remove the CLONE stamp stage, and make the git clone directly
trigger the configure step.