sandbox: don't exit if an app fails
Currently the sandbox follows the same semantic rule as an immutable
pod. As soon as an app fails, it quits immediately.

This behavior does not hold for mutable sandboxes. As soon as an app
fails the sandbox needs to stay alive but track the exit code of the
failed app appropriately.

Also in case of restarts the /rkt/status files need to be readjusted.

This fixes it.

Also fixes #3472