common/overlay: allow data directory name with colon character
This patch provides an sanitizing function used to escape the colon
characters in the directory names of the mount command.

According to the Linux documentation
https://www.kernel.org/doc/Documentation/filesystems/overlayfs.txt it
is used as a separator symbol in case of multi lower-layer mount.

Separates the mount options rendering, thus make it possible to write
pure unit tests for common/overlay package. Functional tests, that
validates common/overlay.Mount function moved to the "tests"
sub-directory.