tests/rkt_stop: Wait for 'stop' command to complete
Do not just 'fire-and-forget' the issuing of the 'stop' command,
but wait for it to complete, or fail.
From testing, it would appear that not waiting for the command to
complete may be responsible for the spurious hangups seen during
CI testing, which may be attributed to not draining the tty of
the command via gexpect.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>