travis: remove "gimme.local" script
This patch removes a "gimme.local" script used within the Travis CI
in order to install the correct version of the Go compiler for aarch64
architecture as the version for Travis was broken.

An issue was fixed in the Travis CI itself, therefore the copy of
"gimme.local" is no longer required within rkt source tree. Updated
the configuration of Travis CI accordingly.