tests: fix when $HOME is only accessible by current user
Previously tested by executing rkt binary with user "nobody". This fails
when the $HOME folder of the user is only accessible by the user itself.
Now the tests first choose the owner of the rkt binary file and if that
is root they use the "nobody" user as a fallback.