stage1/kvm: Change RAM calculation, and increase minimum
The current values and algorithm for working out the VM memory
mean that, for cases such as empty pods, we could end up with as
little as 128Mi of RAM. This is too little, and can impact some
startup functions (init/systemd) as the kernel sets some resources
(such as max-threads) according to how much memory the system has.

Improve the situation by setting a minimum RAM size we need,
and adjusting the calculations accordingly.

Modify the relevant unit tests to match, and extend slightly.

Other potential more dynamic solutions (hotplug memory, ballooning
solutions etc. ) are much more complex, and will be investigated for
the future.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>