build / stage1: Transfer xattr data
Currently there is a combination where both xattr data cannot be
transferred from a squashfs image to the filesystem combined with the
fact that it's not there.  This at least enables the transfer in the
future, and silences some of the warnings in the build process.

If it is decided that xattrs should not be explicitly transferred in the
future this should be changed to -no-xattrs.