stage1-usr: Support local systemd source for offline builds
Local systemd source trees are useful here for packagers to build
usr_from_src without requiring network access.  This differentiates
between local and remote systemd sources by checking if the value
from ./configure contains "://".  If the string is found, the build
will preserve existing behavior.  If not, the value is interpreted
as the source directory path on the build system.